src/laboratory_service/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
djongo==1.3.6
pymongo==3.12.3
PyJWT==2.8.0
django-filter==24.1
celery==5.3.6
redis==5.0.1
drf-yasg==1.21.7
requests==2.31.0
python-dateutil==2.8.2
django-environ==0.11.2
boto3==1.34.34  # For S3/MinIO integration 